feat: Discovery can now be easily customized per game
