+++ Assets/Mirror/Components/Discovery/ServerRequest.cs
@@ -4,9 +4,5 @@ namespace Mirror.Discovery
 {
     public class ServerRequest : MessageBase
     {
-        public long secretHandshake;
-
-        // The server that sent this
-        public IPEndPoint EndPoint { get; set; }
     }
 }
